requirements3.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements3.txt requirements3.in
#
click==6.7                # via pip-tools
first==2.0.1              # via pip-tools
invoke==0.12.2
jinja2==2.9.6
MarkupSafe==1.0           # via jinja2
nose==1.3.7
pip-tools==1.9.0
pyyaml==3.12
requests==2.14.2
six==1.10.0               # via pip-tools
couchdb==1.1
pillow==4.3.0
flask==0.12.2